New EDT issue fix
Had a strange EDT complaint pop up
due to nightcode.customizations/make-skin-map
being called on the EDT.  This was never a
problem until recently, and I assume it
is due to jvm updates.  Would get a null
pointer error IFF make-skin-map
was called ON the EDT.  By default, we
want to push all the UI setup stuff onto
the EDT (since substance will complain).

The work-around - since this is infrequent -
is to wrap make-skin-map in a future and
immediately deref it, thus the setup
happens off the EDT.